[PATCH v2] Get MODULES_VADDR by KERNEL_IMAGE_SIZE for x86_64.
In x86_64, since v2.6.26 the KERNEL_IMAGE_SIZE is changed to 512M, and
accordingly the MODULES_VADDR is changed to 0xffffffffa0000000. Before
that, KERNEL_IMAGE_SIZE is 128M, and MODULES_VADDR is 0xffffffff88000000.

However, in v3.12 Kees Cook introduced kaslr to randomise the location
of kernel. And the kernel text mapping addr space is enlarged from 512M
to 1G. That means now KERNEL_IMAGE_SIZE is variable, its value is 512M
with kaslr support not compiled in and 1G with kaslr support compiled
in. Accordingly the MODULES_VADDR is changed too to be:

  #define MODULES_VADDR    (__START_KERNEL_map + KERNEL_IMAGE_SIZE)

So when kaslr is compiled in and enabled, the kernel text mapping addr
space and modules vaddr space need be adjusted. Otherwise makedumpfile
will collapse since the addr for some symbols is not correct.

Hence KERNEL_IMAGE_SIZE need be exported to vmcoreinfo and got in
makedumpfile to help calculate MODULES_VADDR.

v1->v2:
    Fix a code bug Atsushi found.

Signed-off-by: Baoquan He <bhe@redhat.com>