[PATCH v4 2/4] Make the incomplete dumpfile generated by ENOSPC error analyzable.
Since the incomplete dumpfile generated by ENOSPC error can't be anylyzed
by crash utility, but sometimes this file may contain important information
and the panic problem won't be reproduced, then we came up with an idea to
modify the exist data of the incomplete dumpfile to make it analyzable by
crash utility. And each of those dumpfiles has a flag to indicate that it
has been modified. As there are two formats of these dumpfiles, different
methods and flags are needed to deal with each of them,

elf:
This format use the "e_flags" of "elf header" to indicate that it has been
modified.

The method of dealing with "kdump-compressed" format dumpfiles is
implemented
in the next patch.

Signed-of-by: Wang Xiao <wangx.fnst@cn.fujitsu.com>