[PATCH] Bring get_kcore_dump_loads() backward.
Usually pt_loads[] are initialized with PT_LOADs of /proc/vmcore
but it's impossible in --mem-usage since /proc/kcore doesn't have
VtoP mapping in its PT_LOAD, this is why get_kcore_dump_loads()
complements them by vaddr_to_paddr().

pt_loads[] is necessary to read ELF file, so it should be initialized
as early as possible, but it depends on some VMCOREINFO. So right
after getting VMCOREINFO is the best chance to do it.

Signed-off-by: Atsushi Kumagai <kumagai-atsushi@mxc.nes.nec.co.jp>