[PATCH 1/2] Modify the description of the dump level.
There are two kinds of dump level related to cache pages as:

  - Exclude cache pages without private flag (level:2)
  - Exclude all cache pages (level:4)

The former excludes only the cache pages which don't have private
flag while the latter excludes all cache pages.

However, the man page occasionally make users misunderstood due to
the incomplete description, so it would be better to add the
description of each base dump level and to unify the expression.

Signed-off-by: Atsushi Kumagai <kumagai-atsushi@mxc.nes.nec.co.jp>