[PATCH 2/2] Change the part of the pfn counter for page cache.
The current counter (pfn_cache and pfn_cache_private) are
used for different dump level. A non-private cache page will
be counted in pfn_cache when the dump level is 2 while it will
be counted in pfn_cache_private when the level is 4.

Rather than that, it would be better to use the counters for
different *type* of excluded pages.

Signed-off-by: Atsushi Kumagai <kumagai-atsushi@mxc.nes.nec.co.jp>