[PATCH] Optimise the function reserve_diskspace.
In current implementation, reserve_diskspace use the buffer as large as
the space that needs to be reserved. The space is reserved for PT_LOAD.
Theoretically the size of the whole PT_LOAD can be very large. So it's
necessary to make reserve_diskspace use a limited size of memory.

Signed-off-by: Zhou wenjian <zhouwj-fnst@cn.fujitsu.com>