[PATCH v2] Fix a bug in generating incomplete kdump core.
In current implementation, there is a bug in generating incomplete
kdump core.
The current implementation behaves as the following:

1. Write page header into the buffer cd_header.
2. If cd_header is full, flush the cd_header.
3. Write page data into the buffer cd_page.
4. If cd_page is full, flush the cd_page.

When ENOSPC occurs in flushing cd_page, there's still data left in
cd_header. We cannot read page data corresponding to the page
headers in the cd_header even if the page data has been flushed into
disk. The size of page data is 170 times larger than the size of
page header per a single page frame. The difference becomes bigger
if we use compression.

The patch fixs the bug by changing the logic and the new logic is:

1. Before writing page header and data into buffers, if either of the
   buffers will be full, write the data of the buffer into file.
2. When enospc occurs in writing the cd_header into file, fill the
   cd_header with zero and re-write the cd_header.
3. When enospc occurs in writing the cd_page into file, fill part
   of the cd_header with zero according to how many pages in cd_page
   have been written.
4. Then, write the page header and data info buffers.

Signed-off-by: Zhou Wenjian <zhouwj-fnst@cn.fujitsu.com>