[PATCH v2] Code changes to satisfy the coverity scan.
Coverity scan found some uncheked return value when calling
function. So the return value of the function set_page_size
and lseek should be checked to satisfy the coverity scan.

In the function read_sadump_header, block_size is defined as an
unsigned long value, and it will never be less than zero, so this
comparison should be removed.

Signed-off-by: Chao Fan <cfan@redhat.com>