[PATCH 1/2] sadump: Change bit order.
sadump formats associate each bit in a bitmap with a physical
page in reverse order with the kdump-compressed format.

We had not detected this bug for considerably long term because
bitmaps in sadump formats consist mostly of 0x00 and 0xff excluding a
very limited amount of memory space for firmware.

Signed-off-by: Takao Indoh <indou.takao@jp.fujitsu.com>
Signed-off-by: HATAYAMA Daisuke <d.hatayama@jp.fujitsu.com>