[PATCH 2/4] ARM64: Decide VA Bits on the basis of _stext symbol value.
Currently we assume that there are only two possible configuration
supported by kernel.
1) Page Table Level:2, Page Size 64K and VA Bits 42
1) Page Table Level:3, Page Size 4K and VA Bits 39
Ideally, we should have some mechanism to decide these from kernel
symbols, but we have limited symbols in vmcore,  and we can not do much.
So until some one comes with a better way, we use _stext symbol value
for VA bits determination.
We also assume that when VA Bits are 39, Page table level will be 3 and
Page size will be 4K, while when VA Bits are 42, Page table level will
be 2 an Page size will be 64K.

Signed-off-by: Pratyush Anand <panand@redhat.com>