[PATCH] Looking for page.compound_order/compound_dtor to exclude hugepages
* Required for kernel 4.4

Due to some changes in struct page, hugepages wouldn't be removed on
linux 4.4. makedumpfile reads page.lru.prev to get "order" (number of hugepages)
and page.lru.next to get "dtor" (destructor for hugepages) to detect hugepages,
but the offsets of the two was changed in linux 4.4.

  kernel version |      where is order       |       where is dtor
 ----------------+---------------------------+---------------------------
        - v3.19  |         lru.prev          |          lru.next
   v4.0 - v4.3   | compound_order(=lru.prev) | compound_dtor(=lru.next)
   v4.4 -        | compound_order            | compound_dtor

As above, OFFSET(page.compound_order) and OFFSET(page.compound_dtor) are
definitely necessary in VMCOREINFO on linux 4.4 and later.

Further, the content of page.compound_dtor was changed from direct address
of dtor to the ID of it in linux 4.4.

Signed-off-by: Atsushi Kumagai <ats-kumagai@wm.jp.nec.com>