[PATCH] Fix module init base and size offset for kernel v4.5
* Required for kernel 4.5

This is a follow-up patch on the patch provided in post:

"[PATCH] makedumpfile: fix module core base and size offset for kernel v4.5"
by Pratyush Anand

What is missing there is the changes that were done to the module_init entry
of the struct module. The patch does basically the same as the patch for
module_core adapted to module_init.
Here is the patch:
---
Following commit in kernel v4.5 changed struct module, which affects offset
calculation of core and init layout's base and size.

commit 7523e4dc5057e157212b4741abd6256e03404cf1
Author: Rusty Russell <rusty at rustcorp.com.au>
Date:   Thu Nov 26 09:44:08 2015 +1030

    module: use a structure to encapsulate layout.

This patch fixes makedumpfile code to cater the module_init modification.
---

Signed-off-by: Guenther Hutzl <hutzl@linux.vnet.ibm.com>