[PATCH 1/8] ppc64: fix vtop page translation for 4K pages
* Required for kernel 3.7

PUD index size was changed to 9 from 7 since kernel v3.7, but it was
not updated here. Also, currently L4_MASK is set to 0xfff for kernel
v3.10 and above irrespective of page size but it should be based on
l4 index size which is different for 4k & 64K page sizes. This patch
fixes them.

Signed-off-by: Hari Bathini <hbathini@linux.vnet.ibm.com>