[PATCH 3/8] ppc64: address changes in kernel v4.5
* Required for kernel 4.5

Starting with kernel v4.5, PTE RPN shift value for 4K page size
is increased by one to accommodate page soft dirty tracking and
_PAGE_PRESENT bit is changed. Make the corresponding changes here
and replace all instances of pte_shift with pte_rpn_shift to be
in sync with how it is referred to in the kernel. Also, remove
macro definitions that are no longer used.

Signed-off-by: Hari Bathini <hbathini@linux.vnet.ibm.com>