[PATCH] x86_64: Set zero to phys_base for the kernel older than 2.6.22
Since phys_base was introduced in kernel 2.6.22, it's necessary for
VtoP translation. Unfortunately, the value of it cannot be got without
VtoP, so we estimate it by parsing PT_LOAD headers for now.
(BTW, the value of it will be exported from kernel 4.10)

However, get_phys_base_x86_64() can calculates invalid phys_base
before phys_base was introduced like the case below:

  Type           Offset             VirtAddr           PhysAddr
                 FileSiz            MemSiz              Flags  Align
  NOTE           0x0000000000000190 0x0000000000000000 0x0000000000000000
                 0x0000000000000590 0x0000000000000590         0
  LOAD           0x0000000000000720 0xffffffff80000000 0x0000000000100000    // CONFIG_PHYSICAL_START = 0x100000
                 0x00000000008b2000 0x00000000008b2000  RWE    0

We should always set 0 to phsy_base in that older kernel.

Signed-off-by: Atsushi Kumagai <ats-kumagai@wm.jp.nec.com>