[PATCH v2] Add the --partial-dmesg option to dump only non-cleared dmesg
This option will make --dump-dmesg save only the part of the dmesg
buffer since it was last cleared on the crashed kernel (with
dmesg --clear and such) instead of the whole buffer since boot or
wrap-around. It works on kernels with commit f468908bb55a ("printk:
add clear_idx symbol to vmcoreinfo") merged in v4.6 and otherwise
will default to the regular behavior.

Signed-off-by: Ivan Delalande <colona@arista.com>