PATCH 2/2 Check PG_swapbacked for swap cache pages
When page cache is filtered out (dump level bitmap includes 2 or 4),
makedumpfile checks the PG_swapcache bit, but since kernel commit
6326fec1122cde256bd2a8c63f2606e08e44ce1d (v4.10-rc1~7) this bit is
an alias for PG_owner_priv_1, which is also used by filesystem
code (PG_checked) and Xen (PG_pinned and PG_foreign).

With these kernels, the PG_swapcache flag is valid only if
PG_swapbacked is set. A Linux kernel patch has already been
submitted to export the value of PG_swapbacked in VMCOREINFO.

Since there are released kernels in the wild which do not export the
value, a fallback is implemented. I considered these three situations:

  1. Kernels before v2.6.28-rc1~244:
     PG_swapbacked does not exist, so it must not be checked.
     Instead, check PG_swapcache, which is never overloaded for
     another purpose.

  2. Kernels between v2.6.28-rc1~244 and v4.10-rc1~7:
     It is sufficient to check only PG_swapcache, but PG_swapbacked
     may also be checked (it is always set if PG_swapcache is set).

  3. Kernels since v4.10-rc1~7:
     PG_swapbacked must be checked.

If PG_swapbacked value is known (exported or read from debuginfo),
it is always safe to use it (case 2 or 3). If PG_swapbacked is not
known, it is safe to ignore it for cases 1 and 2, but not 3.
Thankfully, the new value of PG_swapcache (since v4.10-rc1~7) is
less than PG_private (which is known), whereas the old value had
always been greater than PG_private. Moreover, the flags between
PG_private and PG_swapbacked haven't changed since v4.10-rc1~7, so
PG_swapbacked can fall back to PG_private + 6 if unknown.

Without this patch, all Xen dumps are unusable, because PG_pinned is
set for all page table pages.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>